RSPEC-2208 : 'Wildcard imports should not be used' implementation (with RSPEC respect)
