+++ java-checks/src/main/java/org/sonar/java/checks/CheckList.java
@@ -336,7 +336,7 @@ public final class CheckList {
       .add(FileCreateTempFileCheck.class)
       .add(BooleanInversionCheck.class)
       .add(InnerStaticClassesCheck.class)
-      .add(AvoidStarImportCheck.class)
+      .add(WildcardImportsShouldNotBeUsedCheck.class)
       .build();
   }
