--- java-checks/src/main/resources/org/sonar/l10n/java/rules/squid/AvoidStarImportCheck.html
@@ -1,10 +1,0 @@
-<p>Importing all classes from a package or static members from a class leads to tight coupling between packages or classes and might lead to problems when a new version of a library introduces name clashes.</p>
-<p>The following code:</p>
-<pre>
-import java.util.*;  // Non-Compliant
-</pre>
-<p>should be refactored into:</p>
-<pre>
-import java.util.List;  // Compliant
-</pre>
-
