+++ java-checks/src/main/resources/org/sonar/l10n/java/rules/squid/S2208.html
@@ -1,0 +1,21 @@
+<p>Blindly importing all the classes in a package clutters the class namespace and could lead to conflicts between classes in different packages with the same name. On the other hand, specifically listing the necessary classes avoids that problem and makes clear which versions were wanted.</p>
+<h2>Noncompliant Code Example</h2>
+<pre>
+import java.sql.*; // Noncompliant
+import java.util.*; // Noncompliant
+
+private Date date; // Date class exists in java.sql and java.util. Which one is this?
+</pre>
+<h2>Compliant Solution</h2>
+<pre>
+import java.sql.Date;
+import java.util.List;
+import java.util.ArrayList;
+
+private Date date;
+</pre>
+<h2>Exceptions</h2>
+<p>Static imports are ignored by this rule. E.G.</p>
+<pre>
+import static java.lang.Math.*;
+</pre>
